Refactor cloudfront URL signing code
