+++ api/src/tooltool_api/aws.py
@@ -3,13 +3,20 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this
 # file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
+import datetime
 import importlib
 
 import boto
 import boto.s3
 import boto.sqs
+from cryptography.hazmat.backends import default_backend
+from cryptography.hazmat.primitives import hashes
+from cryptography.hazmat.primitives import serialization
+from cryptography.hazmat.primitives.asymmetric import padding
+from botocore.signers import CloudFrontSigner
 
 import tooltool_api.lib.log
+
 
 logger = tooltool_api.lib.log.get_logger(__name__)
 
@@ -53,3 +60,20 @@ class AWS(object):
         # special case for S3, which boto does differently than
         # the other services
         return boto.s3.connect_to_region(region_name=region_name, aws_access_key_id=self.access_key_id, aws_secret_access_key=self.secret_access_key)
+
+    def generate_presigned_cloudfront_url(self, url, expire_time, keypair_id, private_key_string):
+        # From https://boto3.amazonaws.com/v1/documentation/api/latest/reference/services/cloudfront.html#generate-a-signed-url-for-amazon-cloudfront
+        def rsa_signer(message):
+            private_key = serialization.load_pem_private_key(
+                private_key_string,
+                password=None,
+                backend=default_backend()
+            )
+            return private_key.sign(message, padding.PKCS1v15(), hashes.SHA1())
+
+        cloudfront_signer = CloudFrontSigner(keypair_id, rsa_signer)
+
+        # Create a signed url that will be valid until the specfic expiry date
+        # provided using a canned policy.
+        return cloudfront_signer.generate_presigned_url(
+            url, date_less_than=expire_time)
